+++ metadata.rb
@@ -3,7 +3,7 @@ maintainer 'Matt Ray'
 maintainer_email 'matt@chef.io'
 license 'Apache-2.0'
 description 'Installs and configures script for pulling ServiceNow data from Automate and notifying ServiceNow'
-version '0.2.0'
+version '0.3.0'
 chef_version '>= 15.0'
 
 source_url 'https://github.com/mattray/servicenow-task-cookbook'